requirements.txt
# Core
streamlit>=1.26.0
pillow>=9.5.0
requests>=2.31.0
httpx>=0.28.1

# Google / Gemini
google-auth>=2.20.0

# PDF / Image
pdf2image>=1.16.3
pdfplumber>=0.7.6
opencv-python-headless>=4.7.0.72
numpy>=1.24.2

# OCR
pytesseract

# Barcode
pyzbar>=0.1.9
